Service update

+++ SerenityHospital.Business/Services/Implements/ServiceService.cs
@@ -1,6 +1,8 @@
-﻿using AutoMapper;
+﻿using System.Numerics;
+using AutoMapper;
 using SerenityHospital.Business.Dtos.ServiceDtos;
 using SerenityHospital.Business.Exceptions.Common;
+using SerenityHospital.Business.Exceptions.Departments;
 using SerenityHospital.Business.Exceptions.Services;
 using SerenityHospital.Business.Services.Interfaces;
 using SerenityHospital.Core.Entities;
@@ -11,12 +13,14 @@ namespace SerenityHospital.Business.Services.Implements;
 public class ServiceService : IServiceService
 {
     readonly IServiceRepository _repo;
+    readonly IDepartmentRepository _departRepo;
     readonly IMapper _mapper;
 
-    public ServiceService(IServiceRepository repo, IMapper mapper)
+    public ServiceService(IServiceRepository repo, IMapper mapper, IDepartmentRepository departRepo)
     {
         _repo = repo;
         _mapper = mapper;
+        _departRepo = departRepo;
     }
 
     public async Task CreateAsync(ServiceCreateDto dto)
@@ -30,8 +34,10 @@ public class ServiceService : IServiceService
     public async Task DeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<Service>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id,"Departments");
         if (entity is null) throw new NotFoundException<Service>();
+
+        if (entity.Departments.Count() > 0) throw new DepartmentIsNotEmptyException();
 
         _repo.Delete(entity);
         await _repo.SaveAsync();
@@ -41,11 +47,13 @@ public class ServiceService : IServiceService
     {
         if(takeAll)
         {
-            return _mapper.Map<IEnumerable<ServiceListItemDto>>(_repo.GetAll());
+            var entity = _repo.GetAll("Departments");
+            return _mapper.Map<IEnumerable<ServiceListItemDto>>(entity);
         }
         else
         {
-            return _mapper.Map<IEnumerable<ServiceListItemDto>>(_repo.FindAll(s=>s.IsDeleted==false));
+            var entity = _repo.FindAll(s => s.IsDeleted == false);
+            return _mapper.Map<IEnumerable<ServiceListItemDto>>(entity);
         }
     }
 
@@ -55,12 +63,12 @@ public class ServiceService : IServiceService
         Service? entity;
         if(takeAll)
         {
-            entity = await _repo.GetByIdAsync(id);
+            entity = await _repo.GetByIdAsync(id,"Departments");
             if (entity == null) throw new NotFoundException<Service>();
         }
         else
         {
-            entity =await _repo.GetSingleAsync(s=>s.IsDeleted==false && s.Id==id);
+            entity =await _repo.GetSingleAsync(s=>s.IsDeleted==false && s.Id==id,"Departments");
             if (entity == null) throw new NotFoundException<Service>();
         }
 
@@ -70,8 +78,11 @@ public class ServiceService : IServiceService
     public async Task ReverteSoftDeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<Service>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id, "Departments");
         if (entity is null) throw new NotFoundException<Service>();
+
+
+        if (entity.Departments.Count() > 0) throw new DepartmentIsNotEmptyException();
 
         _repo.RevertSoftDelete(entity);
         await _repo.SaveAsync();
@@ -80,8 +91,11 @@ public class ServiceService : IServiceService
     public async Task SoftDeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<Service>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id, "Departments");
         if (entity is null) throw new NotFoundException<Service>();
+
+
+        if (entity.Departments.Count() > 0) throw new DepartmentIsNotEmptyException();
 
         _repo.SoftDelete(entity);
         await _repo.SaveAsync();
@@ -94,6 +108,18 @@ public class ServiceService : IServiceService
         if (entity is null) throw new NotFoundException<Service>();
 
         if (await _repo.IsExistAsync(s => s.Name == dto.Name && s.Id != id)) throw new ServiceNameIsExistException();
+
+        foreach (var itemId in dto.DepartmentIds)
+        {
+            var item = await _departRepo.GetByIdAsync(itemId);
+            if (item is null) throw new NotFoundException<Department>();
+
+            var isDepartmentInOtherService = await _repo.IsExistAsync(s => s.Id != id && s.Departments.Any(d => d.Id == item.Id));
+            if (isDepartmentInOtherService) throw new DepartmentIsInOtherSereviceException();
+
+            entity?.Departments.Add(item);
+        }
+
         _mapper.Map(dto,entity);
         await _repo.SaveAsync();
     }
